LayoutTests/platform/ios-simulator/editing/execCommand/indent-img-twice-expected.txt

Test indenting an image element twice.

Initial state:
| <img>
|   src="../resources/abe.png"

After indenting once:
| <img>
|   src="../resources/abe.png"

After indenting again:
| <img>
|   src="../resources/abe.png"
